feat(EPIC-001): Phase 2c — sov:: arithmetic module, eliminate 8-decimal constants
Add lib_types::tokenomics::sov module as single source of truth for SOV
decimal conversions. All SOV amount calculations now go through:
  sov::atoms(whole_tokens) — convert to 18-decimal atomic units
  sov::to_display(atoms)   — convert to human-readable f64
  sov::SCALE              — raw 10^18 scale factor

Migrated ALL 8-decimal SOV constants:
- SOV_ATOMIC_UNITS: 100_000_000 → sov::SCALE (4 files)
- Welcome bonus: 5_000 * 100_000_000 → sov::atoms(5_000)
- UBI monthly: sov::atoms(1_000)
- Validator rewards: sov::atoms(100)
- PoW budget: sov::atoms(21_000_000)
- Domain fee: sov::atoms(10)
- Treasury allocations: all through sov::atoms()
- Display conversions: / 100_000_000 → sov::to_display()

Type widening (u64 → u128):
- WalletInfo.staked_balance, pending_rewards
- PersistedWalletData.balance
- backup_recovery balance tracking
- marketplace payment amounts
- PoW reward accumulators

42 files changed. 1757 tests pass.

Refs: #2093

+++ lib-identity/examples/dao_wallet_demo.rs
@@ -232,7 +232,7 @@ async fn main() -> Result<()> {
         manager.get_dao_wallets_by_type(false).len()
     );
 
-    let total_dao_funds: u64 = dao_wallets.iter().map(|w| w.balance).sum();
+    let total_dao_funds: u128 = dao_wallets.iter().map(|w| w.balance).sum();
     println!("Total DAO Funds: {} SOV", total_dao_funds);
 
     for dao_wallet in dao_wallets {
